src/requirements.txt
discord.py
pymongo
pycrypto